Getting 'make all' to work cross-platform without messing with PATH.

+++ Makefile
@@ -8,15 +8,6 @@ ALL:
 	make check
 	make install
 
-win:
-	make prepare
-	make rcpp
-	make roxygen
-	make clean
-	make buildwin
-	make check
-	make install
-
 docs:
 	make compile
 	make prepare
@@ -26,7 +17,7 @@ docs:
 compile: inst/ADMB/src/densfuns.cpp inst/ADMB/src/detfuns.cpp inst/ADMB/src/secr.tpl
 	if [ $(shell hostname) == "heton" ]; then cd inst/ADMB/src; admb -O secr.tpl; rm -rfv secr.cpp secr.htp secr.o secr.obj; cd ../bin/linux; rm -rfv secr; mv ../../src/secr ./secr; fi
 	if [ $(shell hostname) == "albatross.mcs.st-and.ac.uk" ]; then cd inst/ADMB/src; admb -O secr.tpl; rm -rfv secr.cpp secr.htp secr.o secr.obj; cd ../bin/mac; rm -rfv secr; mv ../../src/secr ./secr; fi
-	if [ $(shell hostname) == "Zeeba" ]; then cd inst/ADMB/src; admb -O secr.tpl; rm -rfv secr.cpp secr.htp secr.o secr.obj; cd ../bin/windows; rm -rfv secr.exe; mv ../../src/secr.exe ./secr.exe; fi
+	if [ $(shell hostname) == "Zeeba" ]; then export OLDPATH=$$PATH; echo $$OLDPATH; export PATH=/c/admb/bin:/c/MinGW/bin:$$PATH; cd inst/ADMB/src; admb -O secr.tpl; rm -rfv secr.cpp secr.htp secr.o secr.obj; cd ../bin/windows; rm -rfv secr.exe; mv ../../src/secr.exe ./secr.exe; export PATH=$$OLDPATH; fi
 
 prepare:
 	rm -rfv man
@@ -39,13 +30,9 @@ roxygen:
 	R --slave -e "library(roxygen2); roxygenise('.')"
 
 build:
+	if [ $(shell hostname) == "Zeeba" ]; then rm -rfv .Rbuildignore; fi
 	R CMD build .
-	mv admbsecr_1.0.2.tar.gz .Rbuildignore/
-
-buildwin:
-	rm -rfv .Rbuildignore
-	R CMD build .
-	mkdir .Rbuildignore
+	if [ $(shell hostname) == "Zeeba" ]; then mkdir .Rbuildignore; fi
 	mv admbsecr_1.0.2.tar.gz .Rbuildignore/
 
 check:
